before lab

+++ src/utils/math_operations.rs
@@ -25,7 +25,7 @@ pub fn factorial(n: u32) -> u32 {
     result
 }
 
-pub fn mod_inverse(a: u8, m: u8) -> Option<u8> {
+pub fn mod_inverse(a: u32, m: u32) -> Option<u32> {
     for x in 1..m {
         if (a * x) % m == 1 {
             return Some(x);
